Automatically backed up by Learn

+++ db/migrate/001_create_costumes.rb
@@ -3,7 +3,7 @@ class CreateCostumes < ActiveRecord::Migration[4.2]
   def change
     create_table :costumes do |t|
       t.text :name
-      t.integer :price
+      t.float :price
     end
   end
 end
